// packages/frontend/src/app/models/rest-federal-state.ts
export interface RestFederalState {
  OBJECTID: number;
  infizierte: number;
  infizierte_pro_ew: number;
  datum: number;
  einwohner: number;
  zuwachs: number;
  zuwachs_prozent: number;
  GlobalID: string;
  bundesland: string;
  verstorbene: number;
  bundeslandid: number;
  genesene: number;
  positiv: number;
  positiv_pro_ew: number;
  durchschnitt_sieben: number;
  r_eff: null;
  r_eff_lwr: null;
  r_eff_upr: null;
}